requirements.txt
pybullet
jsonschema
numpy
scipy
spatialmath-python
pytest
opencv-python
urdf-parser-py